refactor: drop the conformance kit; the two renderers are independent
The libraries are not required to render identically, so holding them to a
byte-for-byte contract constrained the primary one for no benefit. That became
concrete when dscrbo stopped unrolling nested case classes: the two engines now
differ by design, and every difference had to be argued as an exception to a
contract that no longer described either of them.

Removed: libs/commons/describo-tck, both TckConformanceSuites and their
adapters, both *ParitySuites, the test-scope moduleDeps, and every
byte-identical/parity claim in the two READMEs. Each README now states plainly
that the libraries are independent, with a short table of how they differ.

KnownDivergenceSuite existed only to enumerate exceptions to the contract, but
its assertions cover real behaviour, so rather than delete them each module
keeps them as NamingAndGenericsSuite — the same tests, framed as "this is what
this module does" instead of "this is where it disagrees with the other one".

Coverage was checked rather than assumed before deleting. The naming options
the kit was added for — fullyQualifiedClassName, shortPackagePrefix,
typeNamePrefix/Suffix, fieldNameAndTypeNameSeparator — are all still exercised
in both modules' own suites. Tuples were not: those obligations were describo's
only tuple coverage, so equivalent tests were added to PrintableStructureSuite
before the kit went.

Tests: 555 -> 403, entirely from removing the 74 shared obligations that ran
twice plus the six parity tests. No module lost coverage of its own behaviour.
