--- libs/commons/describo-tck/package.mill
@@ -1,33 +1,0 @@
-package build.libs.commons.`describo-tck`
-
-import mill._
-import mill.javalib._
-import mill.javalib.publish._
-import mill.scalalib.SbtModule
-import mill.scalalib.scalafmt.ScalafmtModule
-
-/** Shared conformance kit for the `describo` and `dscrbo` renderers.
-  *
-  * This module holds the parity specification as data and depends on neither renderer, so `dscrbo`'s zero-dependency
-  * guarantee is untouched: only the two *test* modules depend on this one.
-  */
-object `package` extends SbtModule with PublishModule with ScalafmtModule {
-
-  def scalaVersion = "3.8.4"
-
-  def scalacOptions = Seq("-Wunused:all", "-deprecation", "-feature")
-
-  def mvnDeps = super.mvnDeps() ++ Seq(mvn"org.scalatest::scalatest:3.2.20")
-
-  def pomSettings = PomSettings(
-    description =
-      "Shared conformance test kit pinning byte-identical rendering across the describo and dscrbo toString renderers.",
-    organization = "io.worxbend",
-    url = "https://github.com/worxbend/worxbend",
-    licenses = Seq(License.MIT),
-    versionControl = VersionControl.github("worxbend", "worxbend"),
-    developers = Seq(Developer("limpid-kzonix", "limpid-kzonix", "https://github.com/limpid-kzonix")),
-  )
-
-  def publishVersion = "0.1.0-SNAPSHOT"
-}
