+++ libs/commons/describo/README.md
@@ -7,9 +7,10 @@ hand-written `toString`, no reflection at render time, and no way for a `@Redact
 
 - Scala 3 only (3.8.4).
 - One dependency: `com.softwaremill.magnolia1_3::magnolia`.
-- Sibling module [`dscrbo`](../dscrbo/README.md) produces **byte-identical output** from an inline
-  macro with *zero* dependencies. Pick describo when you want typeclass composition and instances
-  for third-party types; pick dscrbo when you cannot take a dependency.
+- Sibling module [`dscrbo`](../dscrbo/README.md) solves the same problem with an inline macro and
+  *zero* dependencies. The two are **independent and are not required to render identically**. Pick
+  describo when you want typeclass composition and instances for third-party types; pick dscrbo when
+  you cannot take a dependency.
 
 ## Coordinates
 
@@ -294,32 +295,23 @@ value-class parameter cannot omit anything — there would be nothing left to pr
   `com.example.Colour.Red`, which is the more useful spelling. Fixing it here would mean threading the
   parent's name through `split` into every child instance — a change to the `Printable` interface for a
   spelling that only appears under a non-default flag. Both behaviours are pinned by
-  `KnownDivergenceSuite` in each module.
+  `NamingAndGenericsSuite`.
 - **Set and Map iteration order** is the insertion order only up to four elements; beyond that
   Scala switches to a hashed representation. Sort before rendering if you need stable output.
 
-## Parity with `dscrbo`, and the two places it stops
+## How this module differs from `dscrbo`
 
-`describo` and [`dscrbo`](../dscrbo) are two implementations of one specification. For the same input
-and an equivalent `Configuration` they produce **byte-identical** output, and that is enforced rather
-than promised: `libs/commons/describo-tck` holds the specification as data — a catalogue of
-`(fixture, configuration, expected string)` obligations — and both modules run it through a thin
-adapter. Neither *main* module depends on the kit, so `dscrbo`'s zero-dependency guarantee is intact.
+The two are independent implementations and are **not** held to identical output. The differences worth
+knowing about:
 
-Two differences are deliberate and are pinned by `KnownDivergenceSuite` on both sides, so neither can
-quietly become three:
+| | `describo` | `dscrbo` |
+| :-- | :-- | :-- |
+| Nested case class, no instance | auto-derived by magnolia | plain `toString`; nested types are never unrolled |
+| Enum case, qualified name | `com.example.Red` | `com.example.Colour.Red` |
+| Generic case class | derives unaided | needs a `given` for the type argument |
 
-1. **Enum case qualified names.** `describo` prints `com.example.Red`, `dscrbo` prints
-   `com.example.Colour.Red`. See [Known limitations](#known-limitations). Simple names agree.
-2. **Generic case classes.** `describo` derives `Box[Int]` from `Box[A] derives Printable` without
-   ceremony, because its typeclass has real instances for the built-in types a type parameter resolves
-   to. `dscrbo` cannot: its synthesised `derived$Describe[A]` needs a `Describe[A]`, and that module
-   ships no per-type instances by design. **This is a capability `describo` has and `dscrbo` does
-   not.**
-
-Everything else — escaping, `null`, collection and map brackets, `Option`, annotation precedence
-including repeated `@Redacted`, value classes, case objects, sealed families, the multiline threshold
-and every formatting knob — is covered by the shared catalogue and must match exactly.
+Each library pins its own behaviour in its own tests; there is no shared suite and no contract between
+them.
 
 ## Compatibility
 
@@ -355,11 +347,6 @@ made together, on purpose, before the first release:
   reinterpreted in scattered conditionals.
 - Annotation scanning happens once per typeclass instance, at derivation time, and is materialised
   as a `Vector` — not a lazy `View` whose filter re-runs on every `size` and every `map`.
-- Byte-for-byte parity with `dscrbo` is a test, not a promise. `PrintableParitySuite` here and
-  `DescribeParitySuite` there declare the same fourteen-field `TestedType` with the same values and
-  assert the same three strings under the same three configurations. Changing one module's output
-  without changing the other's breaks both suites, which is the point — keep the fixture name, the
-  fields, the values and the expected strings identical, or the comparison silently stops comparing.
 - The build enables `-Wunused:all -deprecation -feature` and mixes in Mill's `ScalafmtModule`, the
   same as `dscrbo`, so dead code and hand-formatting are caught by `./mill libs.commons.describo.compile`
   and `./mill libs.commons.describo.checkFormat` rather than by review.
