+++ libs/commons/describo/package.mill
@@ -28,8 +28,6 @@ object `package` extends SbtModule with PublishModule with ScalafmtModule {
 
   object test extends SbtTests with TestModule.ScalaTest with ScalafmtModule {
 
-    // The conformance kit is test-scope only, so it never reaches this module's published dependencies.
-    def moduleDeps = super.moduleDeps ++ Seq(build.libs.commons.`describo-tck`)
 
     def mvnDeps = super.mvnDeps() ++ Seq(mvn"org.scalatest::scalatest:3.2.20")
