+++ libs/commons/dscrbo/package.mill
@@ -31,8 +31,6 @@ object `package` extends SbtModule with PublishModule with ScalafmtModule { oute
     // captures one `Quotes`. Deliberately kept off the published module so consumers do not pay for the check.
     def scalacOptions = outer.scalacOptions() ++ Seq("-Xcheck-macros")
 
-    // The conformance kit is test-scope only, so it never reaches this module's published dependencies.
-    def moduleDeps = super.moduleDeps ++ Seq(build.libs.commons.`describo-tck`)
 
     def mvnDeps = super.mvnDeps() ++ Seq(mvn"org.scalatest::scalatest:3.2.20")
